backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "numpy>=2.3.5",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "redis[hiredis]>=7.1.0",
    "uvicorn>=0.38.0",
]
